Phase 1.5: FreeType2 text rendering + Linux portability fixes
- Implement FreeType2 + Fontconfig for Linux text rendering (fighter19 pattern)

- Platform isolation: GDI functions wrapped in #ifdef _WIN32

- Portability: w3d_dep.cpp size_t, surfaceclass.cpp uintptr_t, string_compat.cpp _strupr

- Build progress: 3/714 to 217/914 (+214 files)

+++ Core/Libraries/Source/WWVegas/WW3D2/surfaceclass.cpp
@@ -47,6 +47,9 @@
  *   SurfaceClass::Is_Monochrome -- Checks if surface is monochrome or not                     *
  * - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 
+// TheSuperHackers @build fbraz 11/02/2026 Bender - Linux portability: uintptr_t for 64-bit safe pointer arithmetic
+#include <stdint.h>
+
 #include "surfaceclass.h"
 #include "formconv.h"
 #include "dx8wrapper.h"
@@ -575,7 +578,8 @@ void SurfaceClass::FindBB(Vector2i *min,Vector2i*max)
 		for (x = min->I; x < max->I; x++) {
 
 			// HY - this is not endian safe
-			unsigned char *alpha=(unsigned char*) ((unsigned int)lock_rect.pBits+(y-min->J)*lock_rect.Pitch+(x-min->I)*size);
+			// TheSuperHackers @build fbraz 11/02/2026 Bender - 64-bit safe pointer arithmetic
+			unsigned char *alpha=(unsigned char*) ((uintptr_t)lock_rect.pBits+(y-min->J)*lock_rect.Pitch+(x-min->I)*size);
 			unsigned char myalpha=alpha[size-1];
 			myalpha=(myalpha>>(8-alphabits)) & mask;
 			if (myalpha) {
@@ -649,7 +653,8 @@ bool SurfaceClass::Is_Transparent_Column(unsigned int column)
 	for (y = 0; y < (int) sd.Height; y++)
 	{
 		// HY - this is not endian safe
-		unsigned char *alpha=(unsigned char*) ((unsigned int)lock_rect.pBits+y*lock_rect.Pitch);
+		// TheSuperHackers @build fbraz 11/02/2026 Bender - 64-bit safe pointer arithmetic
+		unsigned char *alpha=(unsigned char*) ((uintptr_t)lock_rect.pBits+y*lock_rect.Pitch);
 		unsigned char myalpha=alpha[size-1];
 		myalpha=(myalpha>>(8-alphabits)) & mask;
 		if (myalpha) {
