+++ Core/Libraries/Source/WWVegas/WWSaveLoad/persistfactory.h
@@ -117,12 +117,14 @@ SimplePersistFactoryClass<T,CHUNKID>::Load(ChunkLoadClass & cload) const
 }
 
 
+// TheSuperHackers @build 10/02/2026 Bender
+// Linux 64-bit: uintptr_t instead of uint32 for pointer serialization
 template<class T, int CHUNKID> void
 SimplePersistFactoryClass<T,CHUNKID>::Save(ChunkSaveClass & csave,PersistClass * obj) const
 {
-	uint32 objptr = (uint32)obj;
+	uintptr_t objptr = (uintptr_t)obj;
 	csave.Begin_Chunk(SIMPLEFACTORY_CHUNKID_OBJPOINTER);
-	csave.Write(&objptr,sizeof(uint32));
+	csave.Write(&objptr,sizeof(uintptr_t));
 	csave.End_Chunk();
 
 	csave.Begin_Chunk(SIMPLEFACTORY_CHUNKID_OBJDATA);
